ac_str: Fix some -Wconversion warnings
If compiling with -Wconversion the following was thrown against ac_net.c

ac_str.c: In function ‘ac_str_split’:
ac_str.c:79:43: warning: conversion to ‘long unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
   79 |   fields = realloc(fields, sizeof(char *) * i);
      |                                           ^
ac_str.c:84:42: warning: conversion to ‘long unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
   84 |  fields = realloc(fields, sizeof(char *) * i);
      |                                          ^
ac_str.c: In function ‘ac_str_levenshtein’:
ac_str.c:183:11: warning: conversion from ‘size_t’ {aka ‘long unsigned int’} to ‘int’ may change value [-Wconversion]
  183 |   v0[i] = i;
      |           ^
ac_str.c:188:11: warning: conversion from ‘size_t’ {aka ‘long unsigned int’} to ‘int’ may change value [-Wconversion]
  188 |   v1[0] = i + 1;
      |           ^

They are likely harmless in general, however the first two can be
properly fixed by making i an unsigned int.

The second two we really just silence the warnings...

+++ src/ac_str.c
@@ -56,7 +56,7 @@ char **ac_str_split(const char *string, int delim, int flags)
 	char *p;
 	char **fields;
 	char *tok;
-	int i = 1;
+	unsigned int i = 1;
 
 	/* Check for unknown flags */
 	if (flags & ~(AC_STR_SPLIT_STRICT)) {
@@ -180,12 +180,12 @@ int ac_str_levenshtein(const char *s, const char *t)
 	v1 = malloc((tlen + 1) * sizeof(int));
 
 	for (i = 0; i < tlen + 1; i++)
-		v0[i] = i;
+		v0[i] = (int)i;
 
 	for (i = 0; i < slen; i++) {
 		size_t j;
 
-		v1[0] = i + 1;
+		v1[0] = (int)i + 1;
 
 		for (j = 0; j < tlen; j++) {
 			int cost = (s[i] == t[j]) ? 0 : 1;
